Refactoring the file data to be in the models rather than response package
