+++ app/app_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"github.com/mortedecai/gbb/client"
 	"github.com/mortedecai/gbb/client/mocks"
+	"github.com/mortedecai/gbb/models"
 	"github.com/mortedecai/gbb/response"
 	"go.uber.org/mock/gomock"
 	"go.uber.org/zap"
@@ -133,7 +134,7 @@ func createEmptyDownloadResponse() *http.Response {
 	basicResponse := response.GBBDownloadFilesResponse{
 		Success: true,
 		Data: response.GBBFileDownloadData{
-			Files: make([]response.GBBDownloadFile, 0),
+			Files: make([]models.GBBFileData, 0),
 		},
 	}
 	data, err := json.Marshal(basicResponse)
